Isole le code de la navigation du haut
